requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Database Drivers
# SQLite (included in Python standard library)
# PostgreSQL (uncomment if needed)
# psycopg2-binary==2.9.9
# MySQL (uncomment if needed)
# pymysql==1.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# API Documentation
python-markdown==3.5.1

# Date and Time utilities
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# CORS support
python-cors==1.7.0

# File handling
aiofiles==23.2.1

# Logging
structlog==23.2.0

# Configuration management
pydantic-settings==2.1.0

# Excel/CSV export (optional)
openpyxl==3.1.2
pandas==2.1.3

# HTTP client for external APIs (optional)
requests==2.31.0

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# Monitoring (optional)
prometheus-client==0.19.0

# Development server reload
watchfiles==0.21.0